+++ src/App.css
@@ -1,4 +1,3 @@
 .App {
-  text-align: center;
-  width: 360px;
+  width: 375px;
 }
