+++ src/components/MakerConsole.tsx
@@ -1,11 +1,5 @@
-import Popper from '@mui/material/Popper';
 import React from 'react';
-import { ColorResult, TwitterPicker } from 'react-color';
-import { IconContext } from 'react-icons';
-import { MdOutlineColorLens } from 'react-icons/md';
-import OutsideClickHandler from 'react-outside-click-handler';
 import Select, { ActionMeta, SingleValue } from 'react-select';
-import Switch from 'react-switch';
 import styled from 'styled-components';
 import { setLocalStorage } from '../utils/chromeApi';
 import { IndentOptionsType } from '../utils/types';
@@ -27,13 +21,6 @@ type Props = {
 };
 
 export const MakerConsole: React.FC<Props> = (props) => {
-  const [anchorEl, setAnchorEl] = React.useState<null | SVGElement>(null);
-  const [open, setOpen] = React.useState(false);
-  const handleColorIconClick = (event: React.MouseEvent<SVGElement>) => {
-    setAnchorEl(event.currentTarget);
-    setOpen(!open);
-  };
-
   const handleOnChange = async (
     newValue: SingleValue<{
       value: string;
@@ -60,76 +47,13 @@ export const MakerConsole: React.FC<Props> = (props) => {
     setLocalStorage('scrapboxIndentOption', newIndentOptions);
   };
 
-  const handleColorChange = async ({ hex }: ColorResult) => {
-    if (anchorEl?.id === 'lineColor') {
-      setLocalStorage('scrapboxIndentLineColor', hex);
-      props.setIndentLiningColor(hex);
-    } else if (anchorEl?.id === 'markerColor') {
-      setLocalStorage('scrapboxMarkerColor', hex);
-      props.setMarkerColor(hex);
-    }
-  };
-
-  const handleOnOutsideClick = () => {
-    setOpen(false);
-  };
-
   return (
     <>
       <TitleWrapper>
         <Title>Select Favorite Maker</Title>
-        <SwitchLabel>
-          <Switch
-            checked={props.indentLining}
-            onChange={props.handleIndentLiningChange}
-            onColor="#00b428"
-            boxShadow="0px 1px 5px rgba(0, 0, 0, 0.6)"
-            height={16}
-            width={30}
-            handleDiameter={18}
-            checkedIcon={false}
-            uncheckedIcon={false}
-          />
-          <Spacer width={6} />
-          <span>Indent Visible</span>
-        </SwitchLabel>
-        <IconContext.Provider
-          value={{ size: '20px', style: { padding: '2px' } }}
-        >
-          <ColorLens onClick={handleColorIconClick} id="lineColor" />
-        </IconContext.Provider>
       </TitleWrapper>
 
-      <Popper
-        open={open}
-        anchorEl={anchorEl}
-        placement="bottom-start"
-        modifiers={[
-          {
-            name: 'offset',
-            options: {
-              offset: [10, 10],
-            },
-          },
-        ]}
-      >
-        <OutsideClickHandler onOutsideClick={handleOnOutsideClick}>
-          <TwitterPicker
-            triangle="top-right"
-            onChangeComplete={handleColorChange}
-          />
-        </OutsideClickHandler>
-      </Popper>
-
       <IndentContainer>
-        <MarkerColorIconContainer>
-          <span>Marker Color</span>
-          <IconContext.Provider
-            value={{ size: '20px', style: { padding: '2px' } }}
-          >
-            <ColorLens onClick={handleColorIconClick} id="markerColor" />
-          </IconContext.Provider>
-        </MarkerColorIconContainer>
         {props.indentOptions.map((indentOption) => {
           const spaceNum = +indentOption.label.replace(/[^0-9]/g, '') - 1;
 
@@ -156,46 +80,20 @@ export const MakerConsole: React.FC<Props> = (props) => {
   );
 };
 
-const MarkerColorIconContainer = styled.div`
-  display: flex;
-  align-items: center;
-  position: absolute;
-  right: 6px;
-  top: -13px;
-`;
-
 const TitleWrapper = styled.div`
   display: flex;
   align-items: center;
 `;
 
 const Title = styled.div`
-  font-size: 16px;
+  font-size: 18px;
   text-align: left;
   margin: 10px 12px;
-`;
-
-const ColorLens = styled(MdOutlineColorLens)`
-  color: #00b428;
-  :hover {
-    color: #b4008c;
-    cursor: pointer;
-  }
-`;
-
-const SwitchLabel = styled.label`
-  margin-left: 8px;
-  display: flex;
-  align-items: center;
 `;
 
 const IndentContainer = styled.div`
   margin: 0px 10px;
   position: relative;
-`;
-
-const Spacer = styled.div<{ width: number }>`
-  width: ${(props) => props.width}px;
 `;
 
 const IndentRow = styled.div<{ spaceNum: number }>`
